fea: popups delete

+++ src/app/modules/home/home.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { DataListI } from 'src/app/models/data-list.interface';
 import { HomeService } from 'src/app/services/home.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-home',
@@ -26,8 +27,26 @@ export class HomeComponent implements OnInit {
   }
 
   public removeItem(id : number){
-    this._homeService.removeItem(id).subscribe(console.log)
-    this.listData = this.listData.filter((d:any) => d.id !== id);
+    Swal.fire({
+      title: 'Está seguro?',
+      text: "Desea eliminar el item",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Eliminar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._homeService.removeItem(id).subscribe(console.log)
+        this.listData = this.listData.filter((d:any) => d.id !== id);
+        Swal.fire(
+          'Eliminado!',
+          'El item ha sido eliminado.',
+          'success'
+        )
+      }
+    })
+
   }
 
   public addItem(){
